move hashing code to TxOutConfirmationNumber::new
