+++ transaction/core/src/tx.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2018-2020 MobileCoin Inc.
 
 use alloc::vec::Vec;
+use blake2::digest::Input;
 use core::{
     convert::{TryFrom, TryInto},
     fmt,
@@ -8,7 +9,7 @@ use core::{
 
 use mc_common::{Hash, HashMap};
 use mc_crypto_digestible::Digestible;
-use mc_crypto_keys::{CompressedRistrettoPublic, RistrettoPrivate};
+use mc_crypto_keys::{CompressedRistrettoPublic, RistrettoPrivate, RistrettoPublic};
 use mc_util_repr_bytes::{
     derive_prost_message_from_repr_bytes, typenum::U32, GenericArray, ReprBytes,
 };
@@ -20,6 +21,7 @@ use crate::{
     account_keys::PublicAddress,
     amount::{Amount, AmountError},
     blake2b_256::Blake2b256,
+    domain_separators::TXOUT_CONFIRMATION_NUMBER_DOMAIN_TAG,
     encrypted_fog_hint::EncryptedFogHint,
     onetime_keys::{compute_shared_secret, compute_tx_pubkey, create_onetime_public_key},
     range::Range,
@@ -401,6 +403,14 @@ derive_prost_message_from_repr_bytes!(TxOutMembershipHash);
 pub struct TxOutConfirmationNumber([u8; 32]);
 
 impl TxOutConfirmationNumber {
+    pub fn new(shared_secret: RistrettoPublic) -> Self {
+        let mut hasher = Blake2b256::new();
+        hasher.input(&TXOUT_CONFIRMATION_NUMBER_DOMAIN_TAG);
+        hasher.input(&shared_secret.to_bytes());
+
+        let result: [u8; 32] = hasher.result().into();
+        Self(result)
+    }
     /// Copies self into a new Vec.
     pub fn to_vec(&self) -> Vec<u8> {
         self.0.to_vec()
